misc/HtmlTestSuite/results/test-0113.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Overscroll Behavior: getComputedStyle().overscrollBehavior*</title>
<link href="https://drafts.csswg.org/css-overscroll-behavior/#overscroll-behavior-properties" rel="help"/>
<meta content="overscroll-behavior-x computed value is as specified." name="assert"/>
<meta content="overscroll-behavior-y computed value is as specified." name="assert"/>
<meta content="overscroll-behavior-inline computed value is as specified." name="assert"/>
<meta content="overscroll-behavior-block computed value is as specified." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/computed-testcommon.js"/>
</head>
<body>
<div id="target"/>
<script>
'use strict';
for (let property of ["overscroll-behavior-x", "overscroll-behavior-y", "overscroll-behavior-inline", "overscroll-behavior-block"]) {
  test_computed_value(property, "contain");
  test_computed_value(property, "none");
  test_computed_value(property, "auto");
}
</script>
</body>
</html>
